Add stats to failure accrual
Problem

We do not record stats on FailureAccrual removals (and revivals)

Solution

Modify failure accrual to take a statsReceiver and use that to log
stats.  This highlights the problem that FailureAccrual is essentially a
factory Transformer that can do whatever it wants, so I additionally
deprecated that part of the API because I feel that it's too
open-ended (and certainly impossible to add stats to).  Instead, users
are encouraged to just supply failure accrual params and let us take
care of the transforming.

RB_ID=487093

(sapling split of dad059b8632b7157f64e9b6624766cf7f5c31911)