finagle-core SingletonPool: close underlying
Problem

The SingletonPool failed to close the underlying factory when it
itself was closed. On the face of it, that isn't too bad, except that
if the host went into fail-fast mode BEFORE its factory was closed. In
this case, fail-fast would persist and attempt to reconnect to the
host every 32 seconds.

RB_ID=502817

(sapling split of 090e87ad5848244004913bcc5fc6e241923b5259)