finagle-thrift: Fix EndToEndTest (end-to-end tracing potpourri)
Problem

EndToEndTest was marked flaky. It's actually broken reliably with Stack because of the duplicate ClientAddr annotation. The annotation is added by the ClientDispatcher, so the first request to a service gets double-annotated.

Solution

It's hard to remove the duplicate annotation, so I've changed the test to accept it.

Result

Test passes.

RB_ID=505927

(sapling split of 73602212b5632423eea2bd100102f05474e81f4c)