finagle-core: Introduce the aperture load balancer
The aperture distributor balances load onto a window--the aperture--of
underlying capacity. The distributor exposes a control mechanism so
that a controller can adjust the aperture according to load
conditions.

The window contains a number of discrete serving units, each of which
corresponds to the serving capacity represented by a unit weight. Thus
a single serving unit may be half of a node with weight=2, or two
nodes with weight=0.5. This arrangement allows the aperture
distributor to maintain the weight contract: in aggregate, endpoints
are assigned loads according to their weight and the current load
metric. No load metric is prescribed: this can be mixed in separately.

The underlying nodes are arranged in a consistent fashion: an aperture
of a given size always refers to the same set of nodes; a smaller
aperture to a subset of those nodes. Thus it is relatively harmless to
adjust apertures frequently, since underlying nodes are typically
backed by pools, and will be warm on average.

% csl/smokestack/smoke balancer
p2c.apply		1000000	272 bytes/iter
p2c.update	1000000	46356 bytes/iter
heap.apply	1000000	992 bytes/iter
heap.update	1000000	21437 bytes/iter
aper.apply	1000000	304 bytes/iter
aper.update	1000000	50186 bytes/iter

RB_ID=504485

(sapling split of 7c34eca4a944791ea53a1f0534b19751526e216b)