finagle-http/x: Un-hang EndToEndTest
Problem

When the connection manager observes a chunk appears outside of a
request/response context, it throws an exception, which breaks out
of the Future context because the caller did not catch it. This
results in streamChunks "hanging".

Solution

We catch the exception thrown from ConnectionManager, converting
it into a Throw.

Result

The channel write operation is interrupted correctly, allowing
the Future computation depending on it to correctly handle
failure.

RB_ID=503637

(sapling split of 8828cc408a062dfd04a0dc4ae70e2dc2b0086f9e)