finagle-core: Introduce Status, Service.status, ServiceFactory.status
Problem

Service.isAvailable and ServiceFactory.isAvailable is too
coarse-grained for many uses. For example, we need to be able to
distinguish between a transiently unavailable service and one that is
permanently closed. This is useful in Mux's leasing mechanism as well
as for the upcoming admission controllers.

Solution

Introduce a first-class Status. It's an ADT with variants

	Open
	Busy
	Closed

The Busy variant carries a Future[Unit] indicating when the status has
changed. This is useful for queueing at the appropriate level in the
stack. The other variants don't have arguments.

Discussion

Implementing this without breaking the API was a bit tricky. The
solution is uglier than it should be because of it. I'd like to move
towards finalizing isAvailable, which would make everything much more
consistent. I've inserted notes where appropriate (grep for CSL-1336)
that can be used to track required API changes when we're ready to
pull that trigger.

I've modified the rest of the birdcage, too, to be 'final-safe' -- it
compiles locally if i modify the definition of Service.isAvailable and
ServiceFactory.isAvailable and also fix up the noted usages. The
upshot is that we're very close to being able to do this without
breaking anyone.

This change aims to maintain "isAvailable" parity -- I have not yet
made use of the finer-grained status field; this comes next.

RB_ID=513399

(sapling split of 25515f223152aed829ee41c054a383af5823193c)