Set up SBT for 2.11 builds and remove StreamyJ dependency
Problem

We want to publish 2.11 versions of as many Finagle subprojects as possible.
Some subprojects can't be built for 2.11 at the moment, since they depend on
either Specs or Ostrich, neither of which have builds for 2.11. One subproject
(finagle-example) depends on StreamyJ, which also has no 2.11 build. We're also
using Jackson 2.3.1 internally, and the Jackson Scala module for that version is
not available for 2.11.

Solution

Because StreamyJ is used in a single test and nowhere else in Finagle, I've
simply ported the test to use vanilla Jackson, which allows us to remove the
dependency. The Jackson version used in the SBT builds for Finagle and
twitter-server now depends on the Scala version, and Finagle subprojects that we
can't build for 2.11 are excluded from the root finagle project. I've also
updated the Scala 2.11 versions in the SBT builds for Util and Scrooge from
2.11.2 to 2.11.4.

Result

We can publish 2.11 builds for twitter-server and all but five Finagle
subprojects. Note that finagle-example, finagle-ostrich4, finagle-redis, and
finagle-stress now must be tested and published individually, since they're no
longer aggregated by the root finagle project (this is similar to the current
status of util-eval).

RB_ID=532745

(sapling split of aa9c1eb84c51ebb2ae376537c5990815861094d0)