[finagle-http/x] Fix flaky streaming tests
Problem

Some streaming tests are flaky because of a race between
discarding a Writable and writing to it. The test assumes
that discard happens before the write, causing the write to
expectedly fail. However, sometimes the write happens
before.

Solution

Write repeatedly so that even if we write early, subsequent
writes will come after the discard. This is likely because
subsequent writes wait for antecedent writes to complete.

Result

Passes on 1000 test runs.

RB_ID=541456

(sapling split of 58eb031a817183dee23eae39b354a2daedfdecd2)