finagle-core: Minor fixes to Netty3Transporter.newPipeline
Problem

`InetSocketAddress.getAddress` can return null but this case
was not accounted for in Netty3Transporter. This would lead
to a NullPointerException at runtime if the transporter had
an HttpProxy or SocksProxy configured.

Solution

That case is now accounted for and will no longer NPE.

RB_ID=546310

(sapling split of 22695a5b475d39b4e5fcbfa493dbce1d4227de6e)