finagle-core: Retry forever on WriteExceptions
Problem

WriteExceptions shouldn't have to count against your retry budget,
and should be retried automatically.

Solution

Retrying on WriteExceptions eagerly breaks down without failfast
turned on, and we decide whether to failfast in two places, the
codec and the flag.  I added a RetryingFilter to the stack which
retries 100 times on WriteException, and changed the StackClient
to remove it if it doesn't see a FailFastFactory in the Stack.

Result

We will retry on every finagle "logical" request 100 times on
WriteException.  The RetryingFilters and RetryingPolicies that
people apply should in practice no longer need to account for
WriteExceptions.

RB_ID=540478

(sapling split of eff8363347918e357672710c47e0df2b6f7b6fb7)