[util-core] make Writer extend Closable
All instances of Writer extend the Closable trait
but the java compiler has difficulty understanding
the extension when it occurs as an abstract type
in a return value.

This commit modifies the Writer trait so that it
extends Closable.

RB_ID=551530

(sapling split of 916e918163ff9eebbd1126c0e6528279c6b1cf29)