finagle-mux: Introduce a session failure detector
Problem

We don't have a way to measure session layer liveness. Currently we
rely on connection cutting (which could be due to TCP keepalives, but
they are not really tuned sensibly for our use) and a fail-fast module
to gauge session liveness. This is fickly and unpredictable: we'd like
to extend the notion of liveness in to the application, where it is
much better measured. It's in the same latency and failure domain as
the actual code we're interested in invoking.

Further, TCP keepalives admit little dynamism: we must pick a static
timeout and cannot make any sort of sophisticated or gradual judgement
about an application's state.

Solution

We use Mux's Tping message to transmit periodic messages that are then
used to gauge the peer's liveness. We implement a failure detector,
coined the "Threshold Failure Detector" which dynamically determines
an appropriate timeout based on history.

/**
 * The threshold failure detector uses sessions pings to gauge the health
 * of a peer. It sends ping messages periodically and records their RTTs.
 * These RTTs are averaged over a number of observations.
 *
 * The session is marked [[Status.Busy]] until the first successful ping
 * response has been received.
 *
 * If a ping has been sent and has been outstanding for a time greater
 * than the threshold multiplied by the current exponentially moving
 * average, the session is marked as [[Status.Busy]]. It is marked
 * [[Status.Open]] when the ping message returns.
 *
 * This scheme is pretty conservative, fairly it requires little apriori
 * knowledge: the parameters are used only to tune its sensitivity to
 * history and to bound its failure detection time. The time to detection
 * is bound by the ping period plus the threshold multiplied by the
 * average ping RTT.
*/

Performance notes

We'd like the ping code path to be economical in allocation, since it
scales with the number of sessions (client and server). This change
does a little to improve things, like caching encoded messages, but we
could do a lot more. It may be prudent to push ping processing down to
Netty to gain fine-grained control over allocation. We may also
consider implementing a global rate limit for ping messages. (This
dynamism is permissible in the current threshold detector; it amounts
to adjusting the period dynamically, which presents the trade off of
slower convergence with fewer allocations.) Additionally, we could
dedicate a thread to maintain the health of all Mux sessions, to avoid
extra timer overhead, etc.

RB_ID=542094

(sapling split of 260731bbf835e63e2e23826dc01ee9b734147651)